Changed the navigation bar to be appeared on the every page.
Add going back to login button on the Aboutus page.
